+++ test/time/test_up.c
@@ -8,6 +8,6 @@ int main(void)
 	time_t up = neb_time_up();
 	if (!up)
 		return -1;
-	fprintf(stdout, "uptime: %ld seconds\n", up);
+	fprintf(stdout, "uptime: %lld seconds\n", (long long)up);
 	return 0;
 }
